Fix pointer difference overflow in crash handler memory statistics
UBSan finding 15 (CvGlobals.cpp:2517):
'value -131072 (int) changed to 4294836224 (size_t)' in CustomFilter.

With LAA the application address range is 0x00010000..0xFFFEFFFF. The
byte* subtraction yields ptrdiff_t (32-bit int), and the ~4GB span does
not fit, wrapping to -131073; the subsequent conversion to size_t
happened to restore the intended value. Pointer subtraction whose
result does not fit ptrdiff_t is UB; convert the addresses to size_t
first so the computation is defined. This code only runs inside the
unhandled-exception filter (the session's locals show it fired while
reporting the access violation that ended the run), so there is no
gameplay or determinism impact.

+++ CvGameCoreDLL_Expansion2/CvGlobals.cpp
@@ -2514,7 +2514,8 @@ LONG WINAPI CustomFilter(EXCEPTION_POINTERS* ExceptionInfo)
 		minAddress = (byte*)si.lpMinimumApplicationAddress;
 		maxAddress = (byte*)si.lpMaximumApplicationAddress;
 
-		maxMemory = maxAddress - minAddress + 1;
+		//the full address range exceeds ptrdiff_t on 32 bit - compute in unsigned arithmetic
+		maxMemory = (size_t)maxAddress - (size_t)minAddress + 1;
 
 		byte* currentAddress = minAddress;
